+++ test/cache/lru.cc
@@ -134,7 +134,7 @@ TEST(lru_cache, modifiers)
     cache_type cache(50);
     EXPECT_EQ(cache.size(), 0);
 
-    for (size_t i = 0; i < 50; ++i) {
+    for (int i = 0; i < 50; ++i) {
         cache.insert(i, 2*i);
     }
     EXPECT_EQ(cache.size(), 50);
@@ -199,13 +199,13 @@ TEST(lru_cache, cache_size)
     using cache_type = lru_cache<int, int>;
     cache_type cache(50);
 
-    for (size_t i = 0; i < 50; ++i) {
+    for (int i = 0; i < 50; ++i) {
         cache.insert(i, 2*i);
     }
     EXPECT_EQ(cache.size(), 50);
     EXPECT_EQ(cache.at(0), 0);
 
-    for (size_t i = 50; i < 60; ++i) {
+    for (int i = 50; i < 60; ++i) {
         cache.insert(i, 2*i);
     }
     EXPECT_EQ(cache.size(), 50);
